[KLR] Added project page and styles
Still need to:
• Add functionality for viewing private projects.
• Confirm that the page layout works for 3x4 images in the slideshow
(and possibly make adjustments to make sure images are easy to view at
standard screen sizes).
• Adjust styles for vertically oriented images (e.g. vertical iPhone
layout) so that slideshow fits/looks good on standard screens (since
size is currently determined by width of the container).
• Add some basic link styles.
• A bit more fine tuning.
• Add 'Get Lucky' functionality to nav.
• Ideally adjust carousel to have infinite scroll. (low priority; right
now it slides back to beginning at the end of the scroll, which works
fine).